Debugメッセージ修正・建築量更新を即時処理に
・Debugメッセージをdebugモード以外では表示しないように変更
・建築量更新を即時処理に(1分間の制限も付いてます)
