+++ src/com/github/unchama/listener/MinuteListener.java
@@ -3,7 +3,6 @@ package com.github.unchama.listener;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.github.unchama.task.BuildNumTaskRunnable;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -11,6 +10,7 @@ import org.bukkit.event.Listener;
 import com.github.unchama.event.MinuteEvent;
 import com.github.unchama.gigantic.Gigantic;
 import com.github.unchama.task.AddPotionTaskRunnable;
+import com.github.unchama.task.BuildTaskRunnable;
 import com.github.unchama.task.FlyTaskRunnable;
 import com.github.unchama.task.GiganticSaveTaskRunnable;
 
@@ -80,19 +80,19 @@ public class MinuteListener implements Listener {
 		//1tickにつき1人に処理
 		new FlyTaskRunnable(playerlist).runTaskTimerAsynchronously(plugin, 0, 1);
 	}
-    /**
-     * 建築量
-     *
-     * @param
-     */
-    @EventHandler
-    public void BuildNumListener(MinuteEvent event){
-        List<Player> playerlist = new ArrayList<>(plugin.getServer().getOnlinePlayers());
-
-        if(playerlist.isEmpty()){
-            return;
-        }
-        //1tickにつき1人に処理
-        new BuildNumTaskRunnable(playerlist).runTaskTimerAsynchronously(plugin, 0, 1);
-    }
+	
+	/**
+	 * 1分間の設置量
+	 * 
+	 * @param
+	 */
+	@EventHandler
+	public void BuildNum1minListener(MinuteEvent event){
+		List<Player> playerlist = new ArrayList<Player>(plugin.getServer().getOnlinePlayers());
+		
+		if(playerlist.isEmpty()){
+			return;
+		}
+		new BuildTaskRunnable(playerlist).runTaskTimerAsynchronously(plugin, 0, 1);
+	}
 }
